fix plot_settings
